+++ miosix/filesystem/fat32/diskio.cpp
@@ -49,8 +49,8 @@ DRESULT disk_read (
 	UINT count		/* Number of sectors to read (1..255) */
 )
 {
-    if(pdrv->lseek(sector*512,SEEK_SET)<0) return RES_ERROR;
-    if(pdrv->read(buff,count*512)!=count*512) return RES_ERROR;
+    if(pdrv->lseek(static_cast<off_t>(sector)*512,SEEK_SET)<0) return RES_ERROR;
+    if(pdrv->read(buff,count*512)!=static_cast<ssize_t>(count)*512) return RES_ERROR;
     return RES_OK;
 }
 
@@ -65,8 +65,8 @@ DRESULT disk_write (
 	UINT count		/* Number of sectors to write (1..255) */
 )
 {
-    if(pdrv->lseek(sector*512,SEEK_SET)<0) return RES_ERROR;
-    if(pdrv->write(buff,count*512)!=count*512) return RES_ERROR;
+    if(pdrv->lseek(static_cast<off_t>(sector)*512,SEEK_SET)<0) return RES_ERROR;
+    if(pdrv->write(buff,count*512)!=static_cast<ssize_t>(count)*512) return RES_ERROR;
     return RES_OK;
 }
